requirements.txt
pydub
ruamel.yaml
requests
#genanki
-e git://github.com/mi-ael/genanki#egg=genanki
orderedset
cachier
wanikani-api
jaconv
googledrivedownloader